pyproject.toml
[tool.poetry]
name = "diffcse"
version = "0.1.0"
description = ""
authors = ["GiDaeHyeon <spliter2157@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pytorch-lightning = "^1.7.7"
torch = "^1.12.1"
transformers = "^4.23.1"
torchserve = "^0.6.0"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
